requirements.txt
mistralai>=1.9.10
PyYAML>=6.0.2
ebooklib>=0.18
